// src/types.d.ts
import { type SUPPORTED_LANGUAGES, type AUTO_LENGUAGE} from './constants'

export type Language = keyof typeof SUPPORTED_LANGUAGES
export type AutoLenguage = typeof AUTO_LENGUAGE
export type FromLanguage = Language | AutoLenguage


export interface State {
  result: string,
  fromLanguage: FromLanguage,
  toLanguage: Language,
  fromText: string,
  toText: string,
  loading: boolean,
}

export type Action =
    { type: 'SET_FROM_LANGUAGE', payload: FromLanguage } |
    { type: 'INTERCHANGE_LANGUAGES'} |
    { type: 'SET_TO_LANGUAGE', payload: Language } |
    { type: 'SET_FROM_TEXT', payload: string } |
    { type: 'SET_RESULT', payload: string } |
    { type: 'SET_LOADING', payload: boolean }

export enum SectionType {
  From = 'from',
  To = 'to'
}